atomic-docker/requirements.txt
notion_client
pytest
flask
lancedb
requests
kafka-python
psycopg2-binary
websockets
deepgram-sdk
cryptography
beautifulsoup4
PyAudio
dropbox
pvporcupine
openai
pdfminer.six
tenacity
google-auth-oauthlib
google-api-python-client
simple-salesforce
xero-python
tweepy
newsapi-python
pyowm
python-quickbooks
mailchimp-marketing
linkedin-api
zenpy
greenhouse
yfinance
py-trello
ShopifyAPI
docusign-esign
jira
python-wordpress-xmlrpc
plaid-python
boxsdk
asana
pytest
